Fixing some compiler warnings.
git-svn-id: http://svn.seqan.de/seqan/trunk@10115 e6417c60-b987-48fd-844e-b20f0fcc1017

+++ seqan/core/apps/rabema/do_search.cpp
@@ -23,7 +23,7 @@ int main(int argc, char **argv) {
     Dna5String read = argv[2];
 
     Finder<Dna5String> finder(contig);
-    Pattern<Dna5String, MyersUkkonenReads> pattern(read, -length(read));
+    Pattern<Dna5String, MyersUkkonenReads> pattern(read, -static_cast<int>(length(read)));
 
     EditDistanceScore scoring;
